Simple processor implementation into chrono core and some fiddling with generic types. types might be getting a little out of control (#16)
Simple v1 implementation of a processor
It is very simple and has a lot of problems. So do not worry i dont think it will stay. Just getting the ball rolling
Also messed around with types more. Hopefully types dont end of being more of a liability then helping but still to early to say. We can iterate